refactor: Rename `manifest_dynamic_addresses.rs` to `manifest_address_kinds.rs`
