+++ radix-common/src/data/manifest/model/mod.rs
@@ -1,9 +1,9 @@
 mod manifest_address;
+mod manifest_address_kinds;
 mod manifest_address_reservation;
 mod manifest_blob;
 mod manifest_bucket;
 mod manifest_decimal;
-mod manifest_dynamic_addresses;
 mod manifest_expression;
 mod manifest_non_fungible_local_id;
 mod manifest_precise_decimal;
@@ -11,11 +11,11 @@ mod manifest_proof;
 mod manifest_resource_assertion;
 
 pub use manifest_address::*;
+pub use manifest_address_kinds::*;
 pub use manifest_address_reservation::*;
 pub use manifest_blob::*;
 pub use manifest_bucket::*;
 pub use manifest_decimal::*;
-pub use manifest_dynamic_addresses::*;
 pub use manifest_expression::*;
 pub use manifest_non_fungible_local_id::*;
 pub use manifest_precise_decimal::*;
